This commit reflects the state of tuner as it was used to generate the lineage of versions from 2.2.8alpha to 2.2.8epsilon where alpha is based on the PGNs from a Leorik version that uses material-only evaluation. Epsilon played 609 - 768 - 538  [0.458] against Leorik 2.2 which means the evaluation is ~700 Elo above material-only and ~100 Elo below what could be achieved with Zurichess' dataset. But there's not a plateau reached so the approach to training from scratch may well catch up with a few more iterations.

+++ Leorik.Search/IterativeSearch.cs
@@ -21,7 +21,7 @@ namespace Leorik.Search
         private long _maxNodes;
         private History _history;
         private KillerMoves _killers;
-        private byte _rootRandomness = 0;
+        private int _rootRandomness = 0;
         private Random _rng = new Random();
 
         public long NodesVisited { get; private set; }
@@ -32,7 +32,7 @@ namespace Leorik.Search
         public Span<Move> PrincipalVariation => GetFirstPVfromBuffer(PrincipalVariations, Depth);
 
 
-        public IterativeSearch(BoardState board, byte randomness = 0, long maxNodes = long.MaxValue)
+        public IterativeSearch(BoardState board, int randomness = 0, long maxNodes = long.MaxValue)
         {
             _rootRandomness = randomness;
             _maxNodes = maxNodes;
